move client, df and schema handler to utils
